Phase 21 W1-W3: extract notebook, shell, and validator runners
- task_validation.py: TaskValidator class for declared envs/secrets,
  task contracts, input/return-value validation, and path-type coercion.
- task_runners/shell.py: ShellRunner owns the subprocess registry,
  _run_subprocess, run_logged_command, and the shell driver path.
  Also moves SignalMonitor, ShellTaskError, sanitize_exception, and
  classify_failure here.
- task_runners/notebook.py: NotebookRunner runs NotebookExpr/ScriptExpr
  driver tasks, owns artifact paths, kernel preparation, command
  builders, and manifest extras (incl. cache-hit replay).
- evaluator.py: thinned from 3550 to ~2280 lines; constructs the
  helpers in __post_init__ and delegates from existing dispatch sites.
- tests: monkeypatch targets updated to point at the new runner
  attributes (evaluator._shell_runner._run_subprocess etc.).
