+++ ginkgo/runtime/task_runners/__init__.py
@@ -1,0 +1,28 @@
+"""Task runners that execute the various task kinds.
+
+This package collects the per-kind execution helpers used by
+``_ConcurrentEvaluator``. Splitting them out of ``runtime/evaluator.py``
+keeps each runner small enough to reason about and unit-test in isolation.
+"""
+
+from ginkgo.runtime.task_runners.notebook import (
+    NotebookRunner,
+    NotebookTaskError,
+)
+from ginkgo.runtime.task_runners.shell import (
+    ShellRunner,
+    ShellTaskError,
+    SignalMonitor,
+    classify_failure,
+    sanitize_exception,
+)
+
+__all__ = [
+    "NotebookRunner",
+    "NotebookTaskError",
+    "ShellRunner",
+    "ShellTaskError",
+    "SignalMonitor",
+    "classify_failure",
+    "sanitize_exception",
+]
