+++ tests/test_evaluator.py
@@ -442,7 +442,7 @@ class TestEvaluate:
             raise AssertionError(command)
 
         evaluator = _ConcurrentEvaluator(provenance=recorder, jobs=1, cores=1)
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
         result = evaluator.evaluate(expr)
         manifest = load_manifest(recorder.run_dir)
 
@@ -460,7 +460,7 @@ class TestEvaluate:
         # Re-evaluating with the same notebook hits cache.
         cached = _ConcurrentEvaluator(provenance=recorder, jobs=1, cores=1)
         monkeypatch.setattr(
-            cached,
+            cached._shell_runner,
             "_run_subprocess",
             lambda **_: (_ for _ in ()).throw(AssertionError("cache miss")),
         )
@@ -519,7 +519,7 @@ class TestEvaluate:
             raise AssertionError(command)
 
         first_evaluator = _ConcurrentEvaluator(provenance=first_recorder, jobs=1, cores=1)
-        monkeypatch.setattr(first_evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(first_evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
         first_result = first_evaluator.evaluate(expr)
         first_manifest = load_manifest(first_recorder.run_dir)
         first_html = first_recorder.run_dir / "notebooks" / "task_0000.html"
@@ -539,7 +539,7 @@ class TestEvaluate:
         )
         cached_evaluator = _ConcurrentEvaluator(provenance=second_recorder, jobs=1, cores=1)
         monkeypatch.setattr(
-            cached_evaluator,
+            cached_evaluator._shell_runner,
             "_run_subprocess",
             lambda **_: (_ for _ in ()).throw(AssertionError("cache miss")),
         )
@@ -618,7 +618,7 @@ class TestEvaluate:
             cores=1,
             backend=LocalBackend(pixi_registry=registry),
         )
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
 
         result = evaluator.evaluate(expr)
 
@@ -672,7 +672,7 @@ class TestEvaluate:
             raise AssertionError(command)
 
         evaluator = _ConcurrentEvaluator(provenance=recorder, jobs=1, cores=1, event_bus=bus)
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
 
         result = evaluator.evaluate(expr)
 
@@ -704,7 +704,7 @@ class TestEvaluate:
                 )
             raise AssertionError(command)
 
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
 
         with pytest.raises(RuntimeError, match="ipykernel"):
             evaluator.evaluate(expr)
@@ -783,7 +783,7 @@ class TestEvaluate:
             cores=2,
             backend=LocalBackend(pixi_registry=registry),
         )
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
 
         result = evaluator.evaluate(notebooks)
 
@@ -816,7 +816,7 @@ class TestEvaluate:
             return subprocess.CompletedProcess(args=argv, returncode=0, stdout="run ok", stderr="")
 
         evaluator = _ConcurrentEvaluator(provenance=recorder, jobs=1, cores=1)
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
         result = evaluator.evaluate(expr)
 
         html_path = Path(result)
@@ -842,7 +842,7 @@ class TestEvaluate:
             return subprocess.CompletedProcess(args=argv, returncode=0, stdout="ok", stderr="")
 
         evaluator = _ConcurrentEvaluator(jobs=1, cores=1)
-        monkeypatch.setattr(evaluator, "_run_subprocess", fake_run_subprocess)
+        monkeypatch.setattr(evaluator._shell_runner, "_run_subprocess", fake_run_subprocess)
         result = evaluator.evaluate(expr)
 
         assert Path(result).is_file()
@@ -880,7 +880,7 @@ class TestEvaluate:
 
         ev1 = _ConcurrentEvaluator(jobs=1, cores=1)
         monkeypatch = pytest.MonkeyPatch()
-        monkeypatch.setattr(ev1, "_run_subprocess", fake_subprocess_ok)
+        monkeypatch.setattr(ev1._shell_runner, "_run_subprocess", fake_subprocess_ok)
         try:
             ev1.evaluate(expr_v1)
         finally:
@@ -1251,7 +1251,7 @@ class TestInterruptHandling:
             future_one: (0, "shell"),
             future_two: (1, "python"),
         }
-        evaluator._active_subprocesses = {
+        evaluator._shell_runner._active_subprocesses = {
             tracked_one.pid: tracked_one,  # type: ignore[assignment]
             tracked_two.pid: tracked_two,  # type: ignore[assignment]
         }
@@ -1259,7 +1259,7 @@ class TestInterruptHandling:
         evaluator._python_executor = python_executor
 
         monkeypatch.setattr(
-            evaluator,
+            evaluator._shell_runner,
             "_terminate_subprocess",
             lambda *, process: terminated.append(process.pid),
         )
@@ -1292,17 +1292,17 @@ class TestInterruptHandling:
                 self.returncode = 0
 
             def communicate(self) -> tuple[str, str]:
-                assert 31337 in evaluator._active_subprocesses
+                assert 31337 in evaluator._shell_runner._active_subprocesses
                 return ("stdout", "stderr")
 
-        monkeypatch.setattr("ginkgo.runtime.evaluator.subprocess.Popen", FakePopen)
+        monkeypatch.setattr("ginkgo.runtime.task_runners.shell.subprocess.Popen", FakePopen)
 
-        completed = evaluator._run_subprocess(argv=["echo", "hi"], use_shell=False)
+        completed = evaluator._shell_runner._run_subprocess(argv=["echo", "hi"], use_shell=False)
 
         assert completed.returncode == 0
         assert completed.stdout == "stdout"
         assert completed.stderr == "stderr"
-        assert evaluator._active_subprocesses == {}
+        assert evaluator._shell_runner._active_subprocesses == {}
         assert popen_calls[0][0] == ["echo", "hi"]
 
         if Path("/").anchor == "/":
